@@ -1,14 +1,13 @@
-import React, { useState } from "react";
+import { useState } from "react";
 
 import { useLocalStorage } from "./useLocalStorage";
 
-const TodoContext = React.createContext();
-
-const TodoProvider = (props) => {
+const useTodos = () => {
   // get custom hook useLocalStorage
   const {
     item: todos,
     saveItem: saveTodos,
+    sincronizeItem: sincronizeTodos,
     loading,
     error,
   } = useLocalStorage("TODOS_V1", []);
@@ -61,28 +60,23 @@ const TodoProvider = (props) => {
   };
 
   // Hide ToDos
-  if (todos.completed === false) {
-  }
-  return (
-    <TodoContext.Provider
-      value={{
-        loading,
-        error,
-        totalTodos,
-        completedTodos,
-        searchValue,
-        setSearchValue,
-        searchedTodos,
-        addTodo,
-        completeTodo,
-        deleteTodo,
-        openModal,
-        setOpenModal,
-      }}
-    >
-      {props.children}
-    </TodoContext.Provider>
-  );
+  // if (todos.completed === false) {
+  // }
+  return {
+    loading,
+    error,
+    totalTodos,
+    completedTodos,
+    searchValue,
+    setSearchValue,
+    searchedTodos,
+    addTodo,
+    completeTodo,
+    deleteTodo,
+    openModal,
+    setOpenModal,
+    sincronizeTodos,
+  };
 };
 
-export { TodoContext, TodoProvider };
+export { useTodos };
