@@ -2,7 +2,15 @@ import React from "react";
 
 const EmptyTodos = () => {
   return (
-    <p style={{ textAlign: "center", marginTop: 40 }}>Crea tu primer To Do!</p>
+    <p
+      style={{
+        textAlign: "center",
+        marginTop: 40,
+        backgroundColor: "transparent",
+      }}
+    >
+      Crea tu primer To Do!
+    </p>
   );
 };
 
