@@ -1,16 +1,15 @@
-import React, { useContext } from "react";
-
-// context
-import { TodoContext } from "../TodoContext";
+import React from "react";
 
 // CSS
 import "./TodoCounter.css";
 
-const TodoCounter = () => {
-  const { totalTodos, completedTodos } = useContext(TodoContext);
-
+const TodoCounter = ({ totalTodos, completedTodos, loading }) => {
   return (
-    <div>
+    <div
+      className={`TodoCounter__container ${
+        !!loading && "TodoCounter__container--loading"
+      }`}
+    >
       <h1 className="TodoCounter__title">Tus notas</h1>
       <h2 className="TodoCounter__subtitle">
         Haz completado {completedTodos} de {totalTodos} notas
