@@ -2,7 +2,13 @@ import React from "react";
 
 const TodosError = ({ error }) => {
   return (
-    <p style={{ textAlign: "center", marginTop: 40 }}>
+    <p
+      style={{
+        textAlign: "center",
+        marginTop: 40,
+        backgroundColor: "transparent",
+      }}
+    >
       Desespérate, hubo un error...
     </p>
   );
