@@ -6,7 +6,10 @@ const TodosLoading = () => {
   return (
     <div className="LoadingTodo-container">
       <span className="LoadingTodo-completeIcon"></span>
-      <p style={{ textAlign: "center" }} className="LoadingTodo-Text">
+      <p
+        style={{ textAlign: "center", backgroundColor: "transparent" }}
+        className="LoadingTodo-Text"
+      >
         Cargando TODOs...
       </p>
       <span className="LoadingTodo-deleteIcon"></span>
